# spec/openapi.yaml
openapi: 3.0.0
info:
  description: "A simple Express API"
  title: "Flavors API"
  version: "1.0.0"
servers:
  - url: http://dev.api.com
  - url: http://prod.api.com
tags:
  - name: Flavors
components:
  schemas:
    FlavorItemDto:
      type: object
      required:
        - ingredients
        - name
      properties:
        id:
          type: string
        ingredients:
          type: string
        name:
          type: string
      example:
        id: "123"
        ingredients: vanilla, black cherry
        name: black cherry
    FlavorItemCreateDto:
      type: object
      required:
        - ingredients
        - name
      properties:
        ingredients:
          type: string
        name:
          type: string
      example:
        ingredients: egg yolk, bacon, egg, cheese
        name: baconeggncheese
    FlavorItemUpdateDto:
      type: object
      required:
        - ingredients
        - name
      properties:
        ingredients:
          type: string
        name:
          type: string
      example:
        ingredients: honey, egg yolk, coffee soil
        name: budget breakfast
paths:
  /flavors:
    get:
      operationId: flavorsControllerFindAll
      summary: Read the flavors
      tags:
        - Flavors
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlavorItemDto"
    post:
      operationId: flavorsControllerCreate
      summary: Create a new flavor
      tags:
        - Flavors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlavorItemCreateDto"
      responses:
        400:
          description: Bad Request
          content:
            application/json:
              examples:
                BadRequestException:
                  value:
                    status_code: 400
                    message: string
                    error: Bad Request
        409:
          description: Conflict
          content:
            application/json:
              examples:
                ConflictException:
                  value:
                    status_code: 409
                    message: string
                    error: Conflict
  /flavors/{id}:
    get:
      operationId: flavorsControllerFindOne
      summary: Read the flavor by ID
      tags:
        - Flavors
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlavorItemDto"
        400:
          description: Bad Request
          content:
            application/json:
              examples:
                BadRequestException:
                  value:
                    status_code: 400
                    message: string
                    error: Bad Request
        404:
          description: Not Found
          content:
            application/json:
              examples:
                NotFoundException:
                  value:
                    status_code: 404
                    message: string
                    error: Not Found
    patch:
      operationId: flavorsControllersUpdate
      summary: Update the flavor by ID
      tags:
        - Flavors
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlavorItemUpdateDto"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlavorItemDto"
        400:
          description: Bad Request
          content:
            application/json:
              examples:
                BadRequestException:
                  value:
                    status_code: 400
                    message: string
                    error: Bad Request
        404:
          description: Not Found
          content:
            application/json:
              examples:
                NotFoundException:
                  value:
                    status_code: 404
                    message: string
                    error: Not Found
        409:
          description: Conflict
          content:
            application/json:
              examples:
                ConflictException:
                  value:
                    status_code: 409
                    message: string
                    error: Conflict
    delete:
      operationId: flavorsControllerDelete
      summary: Delete the flavor by ID
      tags:
        - Flavors
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
        204:
          description: No Content
        400:
          description: Bad Request
          content:
            application/json:
              examples:
                BadRequestException:
                  value:
                    status_code: 400
                    message: string
                    error: Bad Request
        404:
          description: Not Found
          content:
            application/json:
              examples:
                NotFoundException:
                  value:
                    status_code: 404
                    message: string
                    error: Not Found